Import of MELO/Redis-1.925 from CPAN.
gitpan-cpan-distribution: Redis
gitpan-cpan-version:      1.925
gitpan-cpan-path:         MELO/Redis-1.925.tar.gz
gitpan-cpan-author:       MELO
gitpan-cpan-maturity:     released

+++ t/tlib/Test/SpawnRedisServer.pm
@@ -4,6 +4,7 @@ package    # Hide from PAUSE
 use strict;
 use warnings;
 use File::Temp;
+use IPC::Cmd qw(can_run);
 use POSIX ":sys_wait_h";
 use base qw( Exporter );
 
@@ -26,8 +27,10 @@ sub redis {
 
   Test::More::diag("Redis port $port, cfg $fn") if $ENV{REDIS_DEBUG};
 
-  ## My local redis PATH
-  $ENV{PATH} = "$ENV{PATH}:/usr/local/redis/sbin";
+  if (! can_run('redis-server')) {
+    Test::More::plan skip_all => "Could not find binary redis-server";
+    return;
+  }
 
   my $c;
   eval { $c = spawn_server($ENV{REDIS_SERVER_PATH} || 'redis-server', $fn) };
@@ -44,8 +47,14 @@ sub spawn_server {
   if ($pid) {    ## Parent
     require Test::More;
     Test::More::diag("Starting server with pid $pid") if $ENV{REDIS_DEBUG};
-    sleep(1);    ## FIXME: we should PING it until he is ready
+
+    ## FIXME: we should PING it until he is ready
+    sleep(1);
+    my $alive = 1;
+
     return sub {
+      return unless $alive;
+
       Test::More::diag("Killing server at $pid") if $ENV{REDIS_DEBUG};
       kill(15, $pid);
 
@@ -59,6 +68,7 @@ sub spawn_server {
         if $ENV{REDIS_DEBUG} && $try > 0;
       unlink('redis-server.log');
       unlink('dump.rdb');
+      $alive = 0;
     };
   }
   elsif (defined $pid) {    ## Child
